Only authorize admin users to put/delete tag (#5024)
* auth only admin users to put/delete tag

* fix typo

* use check admin for delete tag

* move admin auth check to put/delete tag controller

+++ src/rest-server/src/controllers/v2/job.js
@@ -222,6 +222,14 @@ const getSshInfo = asyncHandler(async (req, res) => {
 });
 
 const addTag = asyncHandler(async (req, res) => {
+  // only admin users can add tags
+  if (!req.user.admin) {
+    throw createError(
+      'Unauthorized',
+      'UnauthorizedUserError',
+      'Only admin users are allowed to do this operation.',
+    );
+  }
   await job.addTag(req.params.frameworkName, req.body.value);
   res.status(status('OK')).json({
     status: status('OK'),
@@ -230,6 +238,14 @@ const addTag = asyncHandler(async (req, res) => {
 });
 
 const deleteTag = asyncHandler(async (req, res) => {
+  // only admin users can delete tags
+  if (!req.user.admin) {
+    throw createError(
+      'Unauthorized',
+      'UnauthorizedUserError',
+      'Only admin users are allowed to do this operation.',
+    );
+  }
   await job.deleteTag(req.params.frameworkName, req.body.value);
   res.status(status('OK')).json({
     status: status('OK'),
